requirements.txt
qrcode[pil]
Pillow
PyMuPDF
